--- packages/volto-featured-block/src/components/Blocks/FeaturedBlock/Variations/CarouselFeaturedView.jsx
@@ -1,46 +1,0 @@
-import ConditionalLink from '@plone/volto/components/manage/ConditionalLink/ConditionalLink';
-import React from 'react';
-import { Button } from 'semantic-ui-react';
-
-export const CarouselFeaturedView = (props) => {
-  const { data, isEditMode } = props;
-  return (
-    <div
-      className="banner-top-main"
-      style={{
-        backgroundImage: `url(${data?.url}/@@images/image)`,
-      }}
-    >
-      <div className="ui two column grid">
-        <div className="row">
-          <div className="ten wide column left floated"></div>
-          <div className="six wide column right floated">
-            <div className="banner-top-main-block-title">
-              {data.title && (
-                <h2>
-                  <ConditionalLink
-                    to={data?.href?.[0]?.['@id']}
-                    condition={!isEditMode}
-                  >
-                    {data?.title}
-                  </ConditionalLink>
-                </h2>
-              )}
-              {data.description && <h3>{data?.description} </h3>}
-              {data.hrefText && (
-                <Button
-                  as="a"
-                  size="large"
-                  href={'' + data?.link?.[0]?.['@id']}
-                  primary
-                >
-                  {data.hrefText}
-                </Button>
-              )}
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
