+++ packages/volto-featured-block/src/components/Blocks/FeaturedBlock/Variations/FeaturedViewImageRightRichText.jsx
@@ -1,67 +1,5 @@
-import { TextBlockView } from '@plone/volto-slate/blocks/Text';
-import { DetachedTextBlockEditor } from '@plone/volto-slate/blocks/Text/DetachedTextBlockEditor';
-import ConditionalLink from '@plone/volto/components/manage/ConditionalLink/ConditionalLink';
-import config from '@plone/volto/registry';
-import { Button, Grid, GridColumn, GridRow } from 'semantic-ui-react';
+import { FeaturedBaseView } from './FeaturedBaseView';
+import {  RIGHT } from '../constants';
 export const FeaturedViewImageRightRichText = (props) => {
-  const { data, isEditMode } = props;
-  let Image = config.getComponent('Image').component;
-  return (
-    <Grid divided="vertically">
-      <GridRow>
-        <GridColumn width="6">
-          <div className="featured-block-item-content">
-            {data?.title && (
-              <div className="featured-block-item-content-title">
-                <h2>
-                  <ConditionalLink to={data?.href?.[0]?.['@id']}>
-                    {data?.title}
-                  </ConditionalLink>
-                </h2>
-              </div>
-            )}
-            {data?.description && (
-              <div className="featured-block-item-content-description">
-                {data?.description && <h3>{data.description} </h3>}
-              </div>
-            )}
-            <div className="featured-block-item-content-value">
-              {isEditMode ? (
-                <DetachedTextBlockEditor {...props} />
-              ) : (
-                <TextBlockView {...props} />
-              )}
-            </div>
-
-            {data?.hrefText && (
-              <div className="featured-block-item-content-link">
-                <Button
-                  as="a"
-                  size="large"
-                  href={'' + data?.href?.[0]?.['@id']}
-                  primary
-                >
-                  {data.hrefText}
-                </Button>
-              </div>
-            )}
-          </div>
-        </GridColumn>
-        <GridColumn width="6">
-          {data?.url?.length > 0 && (
-            <Image
-              item={{
-                '@id': data.url,
-                image_field: data.image_field,
-                image_scales: data.image_scales,
-              }}
-              alt=""
-              sizes="50vw"
-              responsive={true}
-            />
-          )}
-        </GridColumn>
-      </GridRow>
-    </Grid>
-  );
+  return <FeaturedBaseView {...props} imagePosition={RIGHT} richText={true} />;
 };
