+++ packages/volto-featured-block/src/components/Blocks/FeaturedBlock/constants.js
@@ -1,0 +1,2 @@
+export const LEFT = 'LEFT';
+export const RIGHT = 'RIGHT';
